Reorganize docker container logic
