+++ mirrored_local_app/constants.py
@@ -1,0 +1,3 @@
+from root import ROOT_PATH
+
+DATA_DUMPER_PATH = str(ROOT_PATH / "local_database" / "DataDumper")
