+++ mirrored_local_app/models/docker_file.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class DockerfileInfo(BaseModel):
+    image_tag: str
+    dockerfile_directory: str | None = None
