+++ mirrored_local_app/timestamp_checker.py
@@ -1,0 +1,28 @@
+import datetime
+import os
+
+
+class TimestampChecker:
+    def __init__(self):
+        self.last_run_time: datetime.datetime | None = self.load_last_run_time()
+
+    def load_last_run_time(self) -> datetime.datetime | None:
+        # Check if file `last_run.txt` exists
+        # If it does, load the last run time
+        if os.path.exists("../local_state/last_run.txt"):
+            with open("../local_state/last_run.txt", "r") as f:
+                return datetime.datetime.strptime(f.read(), "%Y-%m-%d %H:%M:%S")
+        return None
+
+    def last_run_within_24_hours(self):
+        if self.last_run_time is None:
+            return False
+        return datetime.datetime.now() - self.last_run_time < datetime.timedelta(days=1)
+
+    def set_last_run_time(self):
+        # If directory `local_state` doesn't exist, create it
+        if not os.path.exists("../local_state"):
+            os.makedirs("../local_state")
+
+        with open("../local_state/last_run.txt", "w") as f:
+            f.write(datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
