+++ root.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+ROOT_PATH = Path(__file__).parent.resolve()
